+++ packages/dara-components/dara/components/common/table.py
@@ -725,7 +725,8 @@ class Table(ContentComponent):
     :param show_checkboxes: Whether to show or hide checkboxes column when onclick_row is set. Defaults to True
     :param onclick_row: An action handler for when a row is clicked on the table
     :param selected_indices: Optional variable to store the selected rows indices, must be a list of numbers. Note that these indices are
-    the sequential indices of the rows as accepted by `DataFrame.iloc`, not the `row.index` value.
+    the sequential indices of the rows as accepted by `DataFrame.iloc`, not the `row.index` value. If you would like the selection to persist over
+    page reloads, you must set `persist_value=True` on the variable.
     :param search_columns: Optional list defining the columns to be searched, only the columns passed are searchable
     :param searchable: Boolean, if True table can be searched via Input and will only render matching rows
     :param max_rows: if specified, table height will be fixed to accomodate the specified number of rows